Fix memory leak in worker by switching away from `concurrent-ruby`
The `concurrent-ruby` solution exhibited a memory leak. I *think* it was
due to the way I did the re-scheduling after a fetch. A reference was
kept around which prevented GC from cleaning up.

This new approach implements everything on our own, heavily inspired by
the implemenation of Sidekiq regarding this part of the system.

The current implementation does not revive dead runners! Therefore it is
paramount to monitor the number of executable commands.